requirements.txt
openpyxl>=2.4.9
pandas-datareader>=0.5.0
cryptography>=1.8.1
coverage>=4.4.0
numexpr>=2.6.2
scipy>=0.19.0
bottleneck>=1.2.1
bs4>=0.0.1
xlrd>=1.0.0
lxml>=3.7.3
ta-lib>=0.4.16
Keras>=2.1.3


# tushare>=0.7.7
